requirements.txt
Flask
python-Levenshtein
fuzzywuzzy
jellyfish
